1. html pdf ok 2. download ,share, print ok
3. next

atasan query and profile data and update profile data

+++ app/src/main/java/com/coverteam/pta/data/models/InformationDate.java
@@ -2,7 +2,7 @@ package com.coverteam.pta.data.models;
 
 public class InformationDate {
     private String id;
-    private Long jumlahOrangYangCuti;
+    private Integer jumlahOrangYangCuti;
     private String bulan;
     private String tanggal;
     private String tahun;
@@ -10,8 +10,8 @@ public class InformationDate {
     public String getID() { return id; }
     public void setID(String value) { this.id = value; }
 
-    public Long getJumlahOrangYangCuti() { return jumlahOrangYangCuti; }
-    public void setJumlahOrangYangCuti(Long value) { this.jumlahOrangYangCuti = value; }
+    public Integer getJumlahOrangYangCuti() { return jumlahOrangYangCuti; }
+    public void setJumlahOrangYangCuti(Integer value) { this.jumlahOrangYangCuti = value; }
 
     public String getBulan() { return bulan; }
     public void setBulan(String value) { this.bulan = value; }
